backend/requirements.txt
pymupdf==1.23.8
nltk==3.8.1
transformers==4.36.2
torch==2.1.2
pymongo==4.6.1
python-dotenv==1.0.0
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 
fastapi==0.115.9
bcrypt==3.2.0
uvicorn==0.23.1
cloudinary==1.44.0
paddleocr==2.9.1
paddlepaddle==2.6.2
